Refactor into modern namespaced code, with tests outside of the main class
