+++ .gitignore
@@ -1,2 +1,3 @@
 .*
 test.php
+/vendor/
